fix(state): Decrease PoolSize to allow more chains
Maintaining 10 open connections in a pool to the same chain DB results
in a very large number of open connections, and aparently there is a
limit which caps out at about 46 tracked chains, but may vary by system.

Now we use fewer connections in each pool (2), which increases the
tracked chain limit to about 170. A long term sustainable solution is
needed, such as a way to close inactive database connections when not
inuse.

Also improve error reporting and cleanup if interrupted during chain
database creation.

+++ internal/db/fatchain.go
@@ -59,7 +59,8 @@ func NewFATChain(ctx context.Context, dbPath, tokenID string,
 	defer func() {
 		if err != nil {
 			chain.Close()
-			if err := os.Remove(dbPath + chain.DBFile); err != nil {
+			if err := os.Remove(dbPath + chain.DBFile); err != nil &&
+				!os.IsNotExist(err) {
 				chain.Log.Errorf("os.Remove(): %w", err)
 			}
 		}
